cmd/log_mgmt: Fix compilation on linux
Use types as defined in struct cbor_attr_t when passing pointers to
addr union.

Fixes following:
Error: repos/apache-mynewt-mcumgr/cmd/log_mgmt/src/log_mgmt.c:
     In function ‘log_mgmt_show’:
repos/apache-mynewt-mcumgr/cmd/log_mgmt/src/log_mgmt.c:373:29:
     error: initialization of ‘long long int *’ from incompatible
     pointer type ‘int64_t *’ {aka ‘long int *’}
     [-Wincompatible-pointer-types]
  373 |             .addr.integer = &timestamp,
      |                             ^
repos/apache-mynewt-mcumgr/cmd/log_mgmt/src/log_mgmt.c:373:29:
     note: (near initialization for ‘attr[1].addr.integer’)
repos/apache-mynewt-mcumgr/cmd/log_mgmt/src/log_mgmt.c:378:30:
     error: initialization of ‘long long unsigned int *’ from incompatible
     pointer type ‘uint64_t *’ {aka ‘long unsigned int *’}
     [-Wincompatible-pointer-types]
  378 |             .addr.uinteger = &index,
      |                              ^
repos/apache-mynewt-mcumgr/cmd/log_mgmt/src/log_mgmt.c:378:30: note:
     (near initialization for ‘attr[2].addr.uinteger’)

+++ cmd/log_mgmt/src/log_mgmt.c
@@ -353,9 +353,9 @@ log_mgmt_show(struct mgmt_ctxt *ctxt)
     struct log_mgmt_log log;
     CborEncoder logs;
     CborError err;
-    uint64_t index;
+    long long unsigned int index;
     uint32_t next_idx;
-    int64_t timestamp;
+    long long int timestamp;
     int name_len;
     int log_idx;
     int rc;
